Move pdfs inside directories

+++ effingoMoveFiles/Program.cs
@@ -16,8 +16,18 @@ static void moveFiles(string[] main)
         for (int i = 0; i < main.Length; i++)
         {
             //MessageBox.Show(i + " " + Path.GetExtension(main[i]));
-            if(Path.GetExtension(main[i])==".pdf")
+            if(File.Exists(main[i]) && Path.GetExtension(main[i])==".pdf")
                 File.Copy(main[i], Path.Combine(targetFolder, Path.GetFileName(main[i])));
+            else if(Directory.Exists(main[i]))
+            {
+                foreach (string file in Directory.EnumerateFiles(main[i], "*.*", SearchOption.AllDirectories))
+                {
+                    if(Path.GetExtension(file) == ".pdf")
+                    {
+                        File.Copy(file, Path.Combine(targetFolder, Path.GetFileName(file)));
+                    }
+                }
+            }
         }
         //MessageBox.Show(File.ReadAllText("defaultPath.txt"));
     }
